small_api/requirements.txt
Django
django-tastypie
psycopg2
